requirements.txt
asgiref==3.3.4
coverage==5.5
Django==3.2.4
djangorestframework==3.12.4
isort==5.8.0
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
